Rename usages of __core__ to ap_map_flow
