+++ ApertureMapModelTools/data_processing/base_processor.py
@@ -8,7 +8,7 @@ Last Modifed: 2016/10/25
 #
 """
 import os
-from ..__core__ import _get_logger
+from .. import _get_logger
 logger = _get_logger(__name__)
