+++ ApertureMapModelTools/data_processing/histogram.py
@@ -8,7 +8,7 @@ Last Modifed: 2016/10/25
 #
 """
 import scipy as sp
-from ..__core__ import _get_logger, calc_percentile
+from ..ap_map_flow import _get_logger, calc_percentile
 from .base_processor import BaseProcessor
 logger = _get_logger(__name__)
